Create the correct control file when making binary-cross
The correct file needs to be created from the beginning as otherwise
install-msw-dev target, which is a dependency of binary-cross, failed to
install any files into the MSW packages as they were not defined in
debian/control such as it was at the time this target was executed, i.e.
before the correct version could be created in the body of binary-cross
target.

Also add a blank line to control.msw.in to ensure that it is properly
separated from the last package defined in debian/control.in.